+++ docs/CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## [Unreleased]
 
+### Added
+
+- feat: new `callAndPoll` function for an agent to call a canister and poll for the response
+
 ## Changed
 
 - fix: recalculates body to use a fresh `Expiry` when polling for `read_state` requests. This prevents the request from exceeding the `maximum_ingress_expiry` when the replica is slow to respond.
-
-## [2.1.2] - 2024-09-30
-- fix: revert https://github.com/dfinity/agent-js/pull/923 allow option to set agent replica time
 - fix: handle v3 traps correctly, pulling the reject_code and message from the certificate in the error response like v2.
 Example trap error message:
 ```txt
@@ -20,6 +21,9 @@ AgentError: Call failed:
   "Reject message": "foo"
 ```
 - feat: the `UpdateCallRejected` error now exposes  `reject_code: ReplicaRejectCode`, `reject_message: string`, and `error_code?: string` properties directly on the error object.
+
+## [2.1.2] - 2024-09-30
+- fix: revert https://github.com/dfinity/agent-js/pull/923 allow option to set agent replica time
 
 ## [2.1.1] - 2024-09-13